% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_in_na}
\alias{filter_in_na}
\title{Filter In Rows Containing NA in Specified Columns}
\usage{
filter_in_na(.data, ..., if_any_or_all = "if_all")
}
\arguments{
\item{.data}{A dataframe or tibble.}

\item{...}{Columns to check for NA values.}

\item{if_any_or_all}{Should the row be kept if any or all selected columns contain NA?
Must be either "if_any" or "if_all". Defaults to "if_all".}
}
\value{
A tibble with only rows containing NA in specified columns.
}
\description{
This function keeps rows from the data where specified columns contain NA values.
}
