% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_str}
\alias{filter_str}
\title{Filter rows in a data frame based on string matching in a column}
\usage{
filter_str(
  .data,
  col,
  pattern,
  ignore_case = TRUE,
  drop_col = FALSE,
  negate = FALSE,
  na_rm = FALSE
)
}
\arguments{
\item{.data}{A data frame or tibble.}

\item{col}{The column to filter on.}

\item{pattern}{The string pattern to look for.}

\item{ignore_case}{Whether to ignore case (default is TRUE).}

\item{drop_col}{Whether to remove the column that was filtered on (default is FALSE).}

\item{negate}{Whether to keep or remove rows that match the string (default is FALSE).}

\item{na_rm}{Whether to remove NA values (default is FALSE).}
}
\value{
A filtered data frame.
}
\description{
Filter rows in a data frame based on string matching in a column
}
\keyword{internal}
