% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_filter.R
\name{str_filter}
\alias{str_filter}
\title{Filter strings based on a regex pattern}
\usage{
str_filter(string, regex_string, ignore_case = TRUE, negate = FALSE)
}
\arguments{
\item{string}{A character vector to be filtered.}

\item{regex_string}{The regular expression pattern to match against.}

\item{ignore_case}{Logical; should case be ignored in the match? Defaults to TRUE.}

\item{negate}{Logical; should the sense of the match be reversed? Defaults to FALSE.}
}
\value{
A character vector containing the filtered strings.
}
\description{
This function filters a character vector, returning strings that match (or do not match if negate is TRUE) a specified regular expression pattern.
}
\examples{
str_filter(c("apple", "banana", "cherry"), "a")
str_filter(c("apple", "banana", "cherry"), "^a", negate = TRUE)
}
